layout: fix megadrop shadow box
There's a CSS bug that is visible in FF 27.0.1 but not in Chrome Version
33.0.1750.152 under Linux.

The mega-drop div casts a thin line at its top edge, causing a grey
barrier across the entire screen width. I'm by no means a CSS expert,
but tests show that setting the left and bottom box shadow width to
unequal numbers makes the problem disappear.