Dispatch additional ‘editingFinished’ action to indicate, that the user has closed the editor.
Dispatching this action is necessary to clear the editor state after editing is completed. Without this, when reopening the editor it will be initialized with the old values from previous editing until the state updates. Usually this would not be a problem as the components would then also update their mapped state.

But as we use the state to initialize a react-redux form using 'initialValues', the form state will not be updated. So we have to make sure that after the form has opened, it is only set once with the current values.